Add disk-backed db driver and POST code
Big update, introduces a light_types module that defines some base
types. The photons/Sources.py now utilizes this to subclass a Source
type from.

These are connected to the backend through a global package 'backend'
which contains a global variable to instantiate the database driver
into.

As a band-aid, right now, I am manually instantiating a "disk" db driver
which is basically folders-of-JSONs to play with. This is statically
defined in light/__init__.py at the moment, but I hope to move this over
to a dbdriver/dbname in the future.

The big thing here is the beginnings of an object translation layer and
moving data to/from the DB.

I have also created a simple iterator around "get_all" which retrieves
the contents of a data set by constructing a generator to provide these
results, and passes them in a generator API, up the translation chain to
the encapsulating class.