Implement dynamic field identification
Using LightDoc as a base class, identify any subclass's class attributes
where they are a subclass of LightField. If so, attempt some of the
nicer dynamic programming wizardry seen in some ORM systems. This allows
us to move a lot of code out of the subclasses and into the parent
class, which then simply changes behavior slightly based upon subclass
definitions.